updated model fields

+++ CDO/Core/DTOs/Placements/NewPlacement.swift
@@ -14,7 +14,7 @@ struct NewPlacement: Codable {
     let position: String?
     let hireDate: Date?
     let endDate: Date?
-    let daysOnJob: Float?
+    let daysOnJob: Int?
     let day1: String?
     let day2: String?
     let day3: String?
